test lexon.api and lexon.mysql with new parameters to connect to mysql

+++ src/Services/Lexon/Lexon.API/Infrastructure/Services/IUsersService.cs
@@ -19,7 +19,7 @@ namespace Lexon.Infrastructure.Services
 
         Task<List<LexonEntityType>> GetClassificationMasterListAsync();
 
-        Task<long> AddClassificationToListAsync(string idUser, long idCompany, string idMail, long idRelated, short idClassificationType = 1);
+        Task<int> AddClassificationToListAsync(string idUser, long idCompany, string idMail, long idRelated, short idClassificationType = 1);
         Task<long> RemoveClassificationFromListAsync(string idUser, long idCompany, string idMail, long idRelated, short idClassificationType = 1);
 
         //Task<long> AddFileToListAsync(string idUser, long idCompany, long idFile, string nameFile, string descriptionFile = "");
